refactoring - extracting mst from mod - adjusting documentation
